Convert slackevents to TypeScript
