+++ src/lib/utils.ts
@@ -1,0 +1,3 @@
+export const notEmpty = <TValue>(
+  value: TValue | null | undefined
+): value is TValue => value !== null && value !== undefined;
