+++ src/slackevents/.prettierrc.json
@@ -1,0 +1,1 @@
+{}
