+++ src/slackevents/package.json
@@ -1,11 +1,18 @@
 {
   "name": "sample-http",
   "version": "0.0.1",
-  "main": "index.js",
+  "main": "build/slackevents/index.js",
+  "scripts": {
+    "gcp-build": "npm run build",
+    "build": "tsc",
+    "start": "npm run build && npx @google-cloud/functions-framework --target=handler"
+  },
   "dependencies": {
     "firebase-admin": "^8.13.0"
   },
   "devDependencies": {
-    "prettier": "^2.0.5"
+    "@google-cloud/functions-framework": "^3.1.3",
+    "prettier": "^2.8.6",
+    "typescript": "^5.0.2"
   }
 }
